telegram_bot/requirements.txt
python-telegram-bot==20.5
schedule==1.2.0
